Rename ActiveContractsManager
... since it uses the active ledger state, not only active contracts
